Extract glob extensions to own module
We're going to colocate all of the actions, deltas, & services for
a given domain, which will make it easier to reuse them in the future.
