+++ packages/brookjs-cli/src/commands/TestCommand/Check/index.tsx
@@ -2,7 +2,7 @@ import React, { useEffect } from 'react';
 import { useDeltas } from 'brookjs-silt';
 import { unreachable } from 'brookjs-types';
 import { Arguments } from 'yargs';
-import { globLint } from '../../../deltas';
+import * as glob from '../../../glob';
 import { Args } from './types';
 import { reducer } from './reducer';
 import { initialState } from './initialState';
@@ -17,7 +17,7 @@ const Check: React.FC<{ args: Arguments<Args>; rc: unknown; cwd: string }> = ({
   const { state, dispatch } = useDeltas(reducer, initialState(cwd, rc), [exec]);
 
   useEffect(() => {
-    dispatch(globLint.request());
+    dispatch(glob.actions.lint.request());
   }, [dispatch]);
 
   switch (state.status) {
