--- packages/brookjs-cli/src/deltas/index.ts
@@ -1,1 +1,0 @@
-export * from './globDelta';
