+++ packages/brookjs-cli/src/index.ts
@@ -1,4 +1,7 @@
+import * as glob from './glob';
+
 export * from './services';
-export * from './deltas';
 export * from './cli';
 export { default as create } from './create';
+
+export { glob };
