+++ packages/brookjs-cli/src/services/index.ts
@@ -2,7 +2,6 @@ import * as project from './project';
 import * as prettier from './prettier';
 
 export { default as fs } from './fs';
-export { default as glob } from './glob';
 export { project, prettier };
 export { default as ESLintService } from './eslint';
 export { default as WebpackService } from './webpack';
